Moving batch mode to its own tools file
